Bug fix: Add newly-required type conversions in Assert libraries for Solidity 0.8.0

+++ packages/resolver/solidity/AssertInt.sol
@@ -244,7 +244,7 @@ library AssertInt {
             neg = true;
         }
         while (n > 0) {
-            bts[i++] = _utoa(uint8(n % radix)); // Turn it to ascii.
+            bts[i++] = _utoa(uint8(uint(n % radix))); // Turn it to ascii.
             n /= radix;
         }
         // Reverse
